refactor: update operator return types to use int64 and add error handling
This commit refactors the core operator interfaces and their implementations:
1. Change Batch.Do return type from int to int64 for larger count support
2. Update Transformer.Do to return an error alongside existing values
3. Add proper error wrapping in adapter and runner implementations
4. Switch runner to use errgroup for better concurrency error handling
5. Update dependencies and clean up unused go.mod entries
6. Standardize amount/effected return types to int64 across all runners

+++ flow/operator_batch_test.go
@@ -27,12 +27,12 @@ func (b *batch) Prepare() error {
 	return nil
 }
 
-func (b *batch) Do(items []map[string]any) (int, error) {
+func (b *batch) Do(items []map[string]any) (int64, error) {
 	for _, item := range items {
 		_ = item
 	}
 
-	return len(items), nil
+	return int64(len(items)), nil
 }
 
 func (b *batch) BeforeRun() error {
